Feature/flexible remappers (#88)
* Add one-to-one mapper -> Monomapper
* Remapper (one-to-many mapper) is now called Multimapper

---------
